Move to `AdminHome/CloseRequest`
